[Refactor] Nameing folder config to configs
