+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { DatabaseModule } from './database/database.module';
 import { ConfigModule } from '@nestjs/config';
-import { Cors, envValidationObjectSchema, getThrottlerConfig } from './config';
+import { Cors, envValidationObjectSchema, getThrottlerConfig } from './configs';
 import { UserModule } from './user/user.module';
 import { AuthModule } from './auth/auth.module';
 import { AtGuard, RolesGuard } from './auth/guards';
